casi listo

+++ TPE-Grupo13/IntegradorGrupo13/microservicio-viaje/src/main/java/com/grupo13/microservicioviaje/dtos/ReporteViajePeriodoDTO.java
@@ -1,3 +1,12 @@
+/**
+ * 📊 DTO de Respuesta para el Reporte de Viajes por Período.
+ *
+ * Esta clase es un Data Transfer Object (DTO) utilizado para encapsular los
+ * resultados del reporte generado por la consulta JPQL en ViajeRepository.java
+ * (getReporteViajeAnio). Su propósito es transportar de manera estructurada
+ * la información sobre el rendimiento de los monopatines, incluyendo:
+ * ID del monopatín, la cantidad de viajes que realizó en un año específico.
+ */
 package com.grupo13.microservicioviaje.dtos;
 
 import lombok.AllArgsConstructor;
@@ -9,6 +18,6 @@ import lombok.Setter;
 @Setter
 public class ReporteViajePeriodoDTO {
     private String idMonopatin;
-    private int cantidadViajes;
-    private int anio;
+    private Long cantidadViajes;
+    private Integer anio;
 }
